+++ lib/alice.dart
@@ -26,38 +26,50 @@ class Alice {
   /// Icon url for notification
   final String notificationIcon;
 
-  GlobalKey<NavigatorState> _navigatorKey;
-  AliceCore _aliceCore;
-  AliceHttpClientAdapter _httpClientAdapter;
-  AliceHttpAdapter _httpAdapter;
+  ///Max number of calls that are stored in memory. When count is reached, FIFO
+  ///method queue will be used to remove elements.
+  final int maxCallsCount;
+
+  ///Directionality of app. Directionality of the app will be used if set to null.
+  final TextDirection? directionality;
+
+  GlobalKey<NavigatorState>? _navigatorKey;
+  late AliceCore _aliceCore;
+  late AliceHttpClientAdapter _httpClientAdapter;
+  late AliceHttpAdapter _httpAdapter;
 
   /// Creates alice instance.
-  Alice(
-      {GlobalKey<NavigatorState> navigatorKey,
-      this.showNotification = true,
-      this.showInspectorOnShake = false,
-      this.darkTheme = false,
-      this.notificationIcon = "@mipmap/ic_launcher"})
-      : assert(showNotification != null, "showNotification can't be null"),
-        assert(
-            showInspectorOnShake != null, "showInspectorOnShake can't be null"),
-        assert(darkTheme != null, "darkTheme can't be null"),
-        assert(notificationIcon != null, "notificationIcon can't be null") {
+  Alice({
+    GlobalKey<NavigatorState>? navigatorKey,
+    this.showNotification = true,
+    this.showInspectorOnShake = false,
+    this.darkTheme = false,
+    this.notificationIcon = "@mipmap/ic_launcher",
+    this.maxCallsCount = 1000,
+    this.directionality,
+  }) {
     _navigatorKey = navigatorKey ?? GlobalKey<NavigatorState>();
-    _aliceCore = AliceCore(_navigatorKey, showNotification,
-        showInspectorOnShake, darkTheme, notificationIcon);
+    _aliceCore = AliceCore(
+      _navigatorKey,
+      showNotification: showNotification,
+      showInspectorOnShake: showInspectorOnShake,
+      darkTheme: darkTheme,
+      notificationIcon: notificationIcon,
+      maxCallsCount: maxCallsCount,
+      directionality: directionality,
+    );
     _httpClientAdapter = AliceHttpClientAdapter(_aliceCore);
     _httpAdapter = AliceHttpAdapter(_aliceCore);
   }
 
   /// Set custom navigation key. This will help if there's route library.
   void setNavigatorKey(GlobalKey<NavigatorState> navigatorKey) {
-    assert(navigatorKey != null, "navigatorKey can't be null");
-    _aliceCore.setNavigatorKey(navigatorKey);
+    _navigatorKey = navigatorKey;
+    _aliceCore.navigatorKey = navigatorKey;
   }
 
   /// Get currently used navigation key
-  GlobalKey<NavigatorState> getNavigatorKey() {
+  GlobalKey<NavigatorState>? getNavigatorKey() {
     return _navigatorKey;
   }
 
@@ -68,7 +80,6 @@ class Alice {
 
   /// Handle request from HttpClient
   void onHttpClientRequest(HttpClientRequest request, {dynamic body}) {
-    assert(request != null, "httpClientRequest can't be null");
     _httpClientAdapter.onRequest(request, body: body);
   }
 
@@ -76,14 +87,11 @@ class Alice {
   void onHttpClientResponse(
       HttpClientResponse response, HttpClientRequest request,
       {dynamic body}) {
-    assert(response != null, "httpClientResponse can't be null");
-    assert(request != null, "httpClientRequest can't be null");
     _httpClientAdapter.onResponse(response, request, body: body);
   }
 
   /// Handle both request and response from http package
   void onHttpResponse(http.Response response, {dynamic body}) {
-    assert(response != null, "response can't be null");
     _httpAdapter.onResponse(response, body: body);
   }
 
@@ -100,12 +108,8 @@ class Alice {
 
   /// Handle generic http call. Can be used to any http client.
   void addHttpCall(AliceHttpCall aliceHttpCall) {
-    assert(aliceHttpCall != null, "Http call can't be null");
-    assert(aliceHttpCall.id != null, "Http call id can't be null");
     assert(aliceHttpCall.request != null, "Http call request can't be null");
     assert(aliceHttpCall.response != null, "Http call response can't be null");
-    assert(aliceHttpCall.endpoint != null, "Http call endpoint can't be null");
-    assert(aliceHttpCall.server != null, "Http call server can't be null");
     _aliceCore.addCall(aliceHttpCall);
   }
 }