+++ lib/core/alice_core.dart
@@ -6,10 +6,11 @@ import 'package:alice/model/alice_http_error.dart';
 import 'package:alice/model/alice_http_call.dart';
 import 'package:alice/model/alice_http_response.dart';
 import 'package:alice/ui/page/alice_calls_list_screen.dart';
+import 'package:alice/utils/shake_detector.dart';
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:rxdart/rxdart.dart';
-import 'package:shake/shake.dart';
 
 class AliceCore {
   /// Should user be notified with notification if there's new request catched
@@ -30,31 +31,42 @@ class AliceCore {
   /// Icon url for notification
   final String notificationIcon;
 
-  FlutterLocalNotificationsPlugin _flutterLocalNotificationsPlugin;
-  GlobalKey<NavigatorState> _navigatorKey;
+  ///Max number of calls that are stored in memory. When count is reached, FIFO
+  ///method queue will be used to remove elements.
+  final int maxCallsCount;
+
+  ///Directionality of app. If null then directionality of context will be used.
+  final TextDirection? directionality;
+
+  late FlutterLocalNotificationsPlugin _flutterLocalNotificationsPlugin;
+  GlobalKey<NavigatorState>? navigatorKey;
   Brightness _brightness = Brightness.light;
   bool _isInspectorOpened = false;
-  ShakeDetector _shakeDetector;
-  StreamSubscription _callsSubscription;
-  String _notificationMessage;
-  String _notificationMessageShown;
+  ShakeDetector? _shakeDetector;
+  StreamSubscription? _callsSubscription;
+  String? _notificationMessage;
+  String? _notificationMessageShown;
   bool _notificationProcessing = false;
 
   /// Creates alice core instance
-  AliceCore(this._navigatorKey, this.showNotification,
-      this.showInspectorOnShake, this.darkTheme, this.notificationIcon)
-      : assert(showNotification != null, "showNotification can't be null"),
-        assert(
-            showInspectorOnShake != null, "showInspectorOnShake can't be null"),
-        assert(darkTheme != null, "darkTheme can't be null"),
-        assert(notificationIcon != null, "notificationIcon can't be null") {
+  AliceCore(
+    this.navigatorKey, {
+    required this.showNotification,
+    required this.showInspectorOnShake,
+    required this.darkTheme,
+    required this.notificationIcon,
+    required this.maxCallsCount,
+    this.directionality,
+  }) {
     if (showNotification) {
       _initializeNotificationsPlugin();
       _callsSubscription = callsSubject.listen((_) => _onCallsChanged());
     }
     if (showInspectorOnShake) {
       _shakeDetector = ShakeDetector.autoStart(
-        onPhoneShake: () => navigateToCallListScreen(),
+        onPhoneShake: () {
+          navigateToCallListScreen();
+        },
         shakeThresholdGravity: 5,
       );
     }
@@ -93,13 +105,7 @@ class AliceCore {
     }
   }
 
-  /// Set custom navigation key. This will help if there's route library.
-  void setNavigatorKey(GlobalKey<NavigatorState> navigatorKey) {
-    assert(navigatorKey != null, "navigatorKey can't be null");
-    _navigatorKey = navigatorKey;
-  }
-
-  Future<void> _onSelectedNotification(String payload) async {
+  Future<void> _onSelectedNotification(String? payload) async {
     assert(payload != null, "payload can't be null");
     navigateToCallListScreen();
     return;
@@ -126,31 +132,31 @@ class AliceCore {
   }
 
   /// Get context from navigator key. Used to open inspector route.
-  BuildContext getContext() => _navigatorKey?.currentState?.overlay?.context;
+  BuildContext? getContext() => navigatorKey?.currentState?.overlay?.context;
 
   String _getNotificationMessage() {
     final List<AliceHttpCall> calls = callsSubject.value;
     final int successCalls = calls
         .where((call) =>
             call.response != null &&
-            call.response.status >= 200 &&
-            call.response.status < 300)
+            call.response!.status! >= 200 &&
+            call.response!.status! < 300)
         .toList()
         .length;
 
     final int redirectCalls = calls
         .where((call) =>
             call.response != null &&
-            call.response.status >= 300 &&
-            call.response.status < 400)
+            call.response!.status! >= 300 &&
+            call.response!.status! < 400)
         .toList()
         .length;
 
     final int errorCalls = calls
         .where((call) =>
             call.response != null &&
-            call.response.status >= 400 &&
-            call.response.status < 600)
+            call.response!.status! >= 400 &&
+            call.response!.status! < 600)
         .toList()
         .length;
 
@@ -173,25 +179,32 @@ class AliceCore {
     if (errorCalls > 0) {
       notificationsMessage.write("Error: $errorCalls");
     }
-    return notificationsMessage.toString();
+    String notificationMessageString = notificationsMessage.toString();
+    if (notificationMessageString.endsWith(" | ")) {
+      notificationMessageString = notificationMessageString.substring(
+          0, notificationMessageString.length - 3);
+    }
+
+    return notificationMessageString;
   }
 
   Future _showLocalNotification() async {
     _notificationProcessing = true;
     const channelId = "Alice";
     const channelName = "Alice";
-    const channelDescription = "Alice";
     final androidPlatformChannelSpecifics = AndroidNotificationDetails(
-        channelId, channelName, channelDescription,
-        enableVibration: false,
-        playSound: false,
-        largeIcon: DrawableResourceAndroidBitmap(notificationIcon));
+      channelId,
+      channelName,
+      enableVibration: false,
+      playSound: false,
+      largeIcon: DrawableResourceAndroidBitmap(notificationIcon),
+    );
     const iOSPlatformChannelSpecifics =
         IOSNotificationDetails(presentSound: false);
     final platformChannelSpecifics = NotificationDetails(
         android: androidPlatformChannelSpecifics,
         iOS: iOSPlatformChannelSpecifics);
-    final String message = _notificationMessage;
+    final String? message = _notificationMessage;
     await _flutterLocalNotificationsPlugin.show(
         0,
         "Alice (total: ${callsSubject.value.length} requests)",
@@ -205,15 +218,24 @@ class AliceCore {
 
   /// Add alice http call to calls subject
   void addCall(AliceHttpCall call) {
-    assert(call != null, "call can't be null");
-    callsSubject.add([...callsSubject.value, call]);
+    final callsCount = callsSubject.value.length;
+    if (callsCount >= maxCallsCount) {
+      final originalCalls = callsSubject.value;
+      final calls = List<AliceHttpCall>.from(originalCalls);
+      calls.sort(
+          (call1, call2) => call1.createdTime.compareTo(call2.createdTime));
+      final indexToReplace = originalCalls.indexOf(calls.first);
+      originalCalls[indexToReplace] = call;
+
+      callsSubject.add(originalCalls);
+    } else {
+      callsSubject.add([...callsSubject.value, call]);
+    }
   }
 
-  /// Add error to exisng alice http call
+  /// Add error to existing alice http call
   void addError(AliceHttpError error, int requestId) {
-    assert(error != null, "error can't be null");
-    assert(requestId != null, "requestId can't be null");
-    final AliceHttpCall selectedCall = _selectCall(requestId);
+    final AliceHttpCall? selectedCall = _selectCall(requestId);
 
     if (selectedCall == null) {
       AliceUtils.log("Selected call is null");
@@ -226,9 +248,7 @@ class AliceCore {
 
   /// Add response to existing alice http call
   void addResponse(AliceHttpResponse response, int requestId) {
-    assert(response != null, "response can't be null");
-    assert(requestId != null, "requestId can't be null");
-    final AliceHttpCall selectedCall = _selectCall(requestId);
+    final AliceHttpCall? selectedCall = _selectCall(requestId);
 
     if (selectedCall == null) {
       AliceUtils.log("Selected call is null");
@@ -237,19 +257,15 @@ class AliceCore {
     selectedCall.loading = false;
     selectedCall.response = response;
     selectedCall.duration = response.time.millisecondsSinceEpoch -
-        selectedCall.request.time.millisecondsSinceEpoch;
+        selectedCall.request!.time.millisecondsSinceEpoch;
 
     callsSubject.add([...callsSubject.value]);
   }
 
   /// Add alice http call to calls subject
   void addHttpCall(AliceHttpCall aliceHttpCall) {
-    assert(aliceHttpCall != null, "Http call can't be null");
-    assert(aliceHttpCall.id != null, "Http call id can't be null");
     assert(aliceHttpCall.request != null, "Http call request can't be null");
     assert(aliceHttpCall.response != null, "Http call response can't be null");
-    assert(aliceHttpCall.endpoint != null, "Http call endpoint can't be null");
-    assert(aliceHttpCall.server != null, "Http call server can't be null");
     callsSubject.add([...callsSubject.value, aliceHttpCall]);
   }
 
@@ -258,12 +274,11 @@ class AliceCore {
     callsSubject.add([]);
   }
 
-  AliceHttpCall _selectCall(int requestId) => callsSubject.value
-      .firstWhere((call) => call.id == requestId, orElse: () => null);
+  AliceHttpCall? _selectCall(int requestId) =>
+      callsSubject.value.firstWhereOrNull((call) => call.id == requestId);
 
   /// Save all calls to file
   void saveHttpRequests(BuildContext context) {
-    assert(context != null, "context can't be null");
     AliceSaveHelper.saveCalls(context, callsSubject.value, _brightness);
   }
 }