+++ lib/model/alice_http_call.dart
@@ -4,6 +4,7 @@ import 'package:alice/model/alice_http_response.dart';
 
 class AliceHttpCall {
   final int id;
+  late DateTime createdTime;
   String client = "";
   bool loading = true;
   bool secure = false;
@@ -13,12 +14,13 @@ class AliceHttpCall {
   String uri = "";
   int duration = 0;
 
-  AliceHttpRequest request;
-  AliceHttpResponse response;
-  AliceHttpError error;
+  AliceHttpRequest? request;
+  AliceHttpResponse? response;
+  AliceHttpError? error;
 
   AliceHttpCall(this.id) {
     loading = true;
+    createdTime = DateTime.now();
   }
 
   void setResponse(AliceHttpResponse response) {
@@ -30,7 +32,7 @@ class AliceHttpCall {
     var compressed = false;
     var curlCmd = "curl";
     curlCmd += " -X $method";
-    final headers = request.headers;
+    final headers = request!.headers;
     headers.forEach((key, dynamic value) {
       if ("Accept-Encoding" == key && "gzip" == value) {
         compressed = true;
@@ -38,19 +40,19 @@ class AliceHttpCall {
       curlCmd += " -H '$key: $value'";
     });
 
-    final String requestBody = request.body.toString();
-    if (requestBody != null && requestBody != '') {
+    final String requestBody = request!.body.toString();
+    if (requestBody != '') {
       // try to keep to a single line and use a subshell to preserve any line breaks
       curlCmd += " --data \$'${requestBody.replaceAll("\n", "\\n")}'";
     }
 
-    var queryParamMap = request.queryParameters;
+    final queryParamMap = request!.queryParameters;
     int paramCount = queryParamMap.keys.length;
     var queryParams = "";
     if (paramCount > 0) {
       queryParams += "?";
       queryParamMap.forEach((key, dynamic value) {
-        queryParams += '${key}=${value}';
+        queryParams += '$key=$value';
         paramCount -= 1;
         if (paramCount > 0) {
           queryParams += "&";