+++ lib/ui/page/alice_stats_screen.dart
@@ -1,29 +1,31 @@
 import 'package:alice/core/alice_core.dart';
 import 'package:alice/helper/alice_conversion_helper.dart';
 import 'package:alice/model/alice_http_call.dart';
-import 'package:alice/ui/utils/alice_constants.dart';
+import 'package:alice/utils/alice_constants.dart';
 import 'package:flutter/material.dart';
 
 class AliceStatsScreen extends StatelessWidget {
   final AliceCore aliceCore;
 
-  const AliceStatsScreen(this.aliceCore)
-      : assert(aliceCore != null, "aliceCore can't be null");
+  const AliceStatsScreen(this.aliceCore);
 
   @override
   Widget build(BuildContext context) {
-    return Theme(
-      data: ThemeData(
-          brightness: aliceCore.brightness,
-          accentColor: AliceConstants.lightRed),
-      child: Scaffold(
-        appBar: AppBar(
-          title: const Text("Alice - HTTP Inspector - Stats"),
-        ),
-        body: Container(
-          padding: const EdgeInsets.all(8),
-          child: ListView(
-            children: _buildMainListWidgets(),
+    return Directionality(
+      textDirection: aliceCore.directionality ?? Directionality.of(context),
+      child: Theme(
+        data: ThemeData(
+            brightness: aliceCore.brightness,
+            accentColor: AliceConstants.lightRed),
+        child: Scaffold(
+          appBar: AppBar(
+            title: const Text("Alice - HTTP Inspector - Stats"),
+          ),
+          body: Container(
+            padding: const EdgeInsets.all(8),
+            child: ListView(
+              children: _buildMainListWidgets(),
+            ),
           ),
         ),
       ),
@@ -58,8 +60,6 @@ class AliceStatsScreen extends StatelessWidget {
   }
 
   Widget _getRow(String label, String value) {
-    assert(label != null, "label can't be null");
-    assert(value != null, "value can't be null");
     return Row(
       children: <Widget>[
         Text(
@@ -92,24 +92,24 @@ class AliceStatsScreen extends StatelessWidget {
   int _getSuccessRequests() => calls
       .where((call) =>
           call.response != null &&
-          call.response.status >= 200 &&
-          call.response.status < 300)
+          call.response!.status! >= 200 &&
+          call.response!.status! < 300)
       .toList()
       .length;
 
   int _getRedirectionRequests() => calls
       .where((call) =>
           call.response != null &&
-          call.response.status >= 300 &&
-          call.response.status < 400)
+          call.response!.status! >= 300 &&
+          call.response!.status! < 400)
       .toList()
       .length;
 
   int _getErrorRequests() => calls
       .where((call) =>
           call.response != null &&
-          call.response.status >= 400 &&
-          call.response.status < 600)
+          call.response!.status! >= 400 &&
+          call.response!.status! < 600)
       .toList()
       .length;
 
@@ -119,7 +119,7 @@ class AliceStatsScreen extends StatelessWidget {
   int _getBytesSent() {
     int bytes = 0;
     calls.forEach((AliceHttpCall call) {
-      bytes += call.request.size;
+      bytes += call.request!.size;
     });
     return bytes;
   }
@@ -128,7 +128,7 @@ class AliceStatsScreen extends StatelessWidget {
     int bytes = 0;
     calls.forEach((AliceHttpCall call) {
       if (call.response != null) {
-        bytes += call.response.size;
+        bytes += call.response!.size;
       }
     });
     return bytes;